Memory optimizations, message cap removal, HMAC ring seed, and cross-platform fixes
Core memory optimizations (two largest allocators):
- J-UNIWARD: f32 buffers, sequential subbands, CostMapPtr direct writes (~651MB -> ~187MB, 71% reduction)
- Phase 3 geometric recovery: f32 Complex32 FFT, gather-FFT-scatter columns, BluesteinPlan chirp reuse, LRU LLR cache (max 5), early drops (~500-907MB -> ~93MB)

Message cap removal: 1024 -> u16::MAX (65,535 bytes), adaptive STC params
HMAC ring seed: replaced XOR domain separation with HMAC-SHA256, bumped v4 -> v5
iOS: capacity task cancellation, localization fixes (7 new + 6 updated keys, 15 languages), stealth dialogs
Android: capacity job cancellation
Web: generation counter pattern for capacity cancellation
Cross-platform test fixes: f32 FFT bit-pattern test, relaxed roundtrip tolerance

+++ src/stego/armor/selection.rs
@@ -11,7 +11,7 @@ use crate::stego::cost::CostMap;
 use super::embedding::MAX_ARMOR_ZIGZAG;
 
 /// Cost assigned to stable positions (low, uniform cost for permutation compatibility).
-const STABLE_COST: f64 = 1.0;
+const STABLE_COST: f32 = 1.0;
 
 /// Compute a stability map for the given DCT grid.
 ///
